requirements.txt
websockets
click
pyyaml